Move GIRG headers into dedicated include directory
