+++ include/networkit/generators/GeometricInhomogenousGenerator.hpp
