+++ include/networkit/generators/girgs/Generator.hpp
@@ -1,5 +1,5 @@
 /*
- * Generator.h
+ * Generator.hpp
  *
  *  Created on: 03. May 2019
  *      Author: Christopher Weyand <Christopher.Weyand@hpi.de>, Manuel Penschuck <networkit@manuel.jetzt>
@@ -12,7 +12,7 @@
 #define GENERATORS_GIRGS_GENERATOR_H_
 
 #include <vector>
-#include "../../graph/Graph.h"
+#include "../../graph/Graph.hpp"
 
 namespace NetworKit {
 namespace girgs {
