+++ include/networkit/generators/girgs/Node.hpp
