+++ include/networkit/generators/girgs/WeightScaling.hpp
